partial implementation of new BottomSheet design, added notes to few composable
