+++ app/src/main/java/com/savenko/track/presentation/other/windowInfo/WindowInfo.kt
@@ -32,8 +32,8 @@ data class WindowInfo(
     val screenHeight: Dp
 ) {
     sealed class WindowType {
-        object Compact: WindowType()
-        object Medium: WindowType()
-        object Expanded: WindowType()
+        data object Compact: WindowType()
+        data object Medium: WindowType()
+        data object Expanded: WindowType()
     }
 }
