+++ app/src/main/java/com/savenko/track/presentation/screens/core/TrackScreenManager.kt
@@ -1,20 +1,17 @@
 package com.savenko.track.presentation.screens.core
 
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
-import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import com.savenko.track.data.viewmodels.common.TrackScreenManagerViewModel
 
-/*
-Track screen manager is pager that handles navigation across main Track screens : SettingsTrackScreen, MainTrackScreen, StatisticsTrackScreen
-*/
+/**
+ * Track screen manager is pager that handles navigation across main Track screens : SettingsTrackScreen, MainTrackScreen, StatisticsTrackScreen
+ */
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun TrackScreenManager(navHostController: NavHostController, viewModel: TrackScreenManagerViewModel) {
@@ -24,19 +21,13 @@ fun TrackScreenManager(navHostController: NavHostController, viewModel: TrackScr
     LaunchedEffect(pagerState.currentPage) {
         viewModel.setPagerState(pagerState.currentPage)
     }
-
-    Column(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        HorizontalPager(
-            state = pagerState,
-            modifier = Modifier.weight(1f)
-        ) { page ->
-            when (page) {
-                0 -> SettingsTrackScreen(navHostController)
-                1 -> MainTrackScreen()
-                2 -> StatisticsTrackScreen()
-            }
+    HorizontalPager(
+        state = pagerState
+    ) { page ->
+        when (page) {
+            0 -> SettingsTrackScreen(navHostController)
+            1 -> MainTrackScreen()
+            2 -> StatisticsTrackScreen()
         }
     }
 }
